backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4

chromadb==0.4.24

langchain==0.1.20
langchain-community==0.0.40
langchain-text-splitters==0.0.2
langchain-openai==0.1.7
langchain-huggingface==0.1.0
langchain-chroma==0.1.2
langchain-ollama==0.1.0

sentence-transformers==2.7.0
pypdf==4.0.2
